ai-service/RAG/requirements.txt
# RAG 시스템 필수 패키지
requests>=2.31.0
numpy>=1.24.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# 선택적 패키지 (성능 향상)
torch>=2.0.0
transformers>=4.30.0

# 유틸리티
python-dotenv>=1.0.0